Rename the whole thing
