+++ .github/workflows/continuous-integration.yaml
@@ -33,7 +33,7 @@ jobs:
     runs-on: ubuntu-20.04
     strategy:
       matrix:
-        workdir: [".", "./cmd/otel-allocator", "./cmd/remote-configuration"]
+        workdir: [".", "./cmd/otel-allocator", "./cmd/operator-opamp-bridge"]
     steps:
     - name: Set up Go
       uses: actions/setup-go@v3
