+++ cmd/operator-opamp-bridge/agent/agent.go
@@ -21,10 +21,10 @@ import (
 
 	"gopkg.in/yaml.v3"
 
-	"github.com/open-telemetry/opentelemetry-operator/cmd/remote-configuration/metrics"
-	"github.com/open-telemetry/opentelemetry-operator/cmd/remote-configuration/operator"
+	"github.com/open-telemetry/opentelemetry-operator/cmd/operator-opamp-bridge/metrics"
+	"github.com/open-telemetry/opentelemetry-operator/cmd/operator-opamp-bridge/operator"
 
-	"github.com/open-telemetry/opentelemetry-operator/cmd/remote-configuration/config"
+	"github.com/open-telemetry/opentelemetry-operator/cmd/operator-opamp-bridge/config"
 
 	"github.com/oklog/ulid/v2"
 	"go.uber.org/multierr"
