+++ cmd/operator-opamp-bridge/agent/collector_key.go
