+++ cmd/operator-opamp-bridge/agent/collector_key_test.go
