+++ cmd/operator-opamp-bridge/agent/testdata/agent.yaml
