+++ cmd/operator-opamp-bridge/agent/testdata/agentbasiccomponentsallowed.yaml
