+++ cmd/operator-opamp-bridge/agent/testdata/agentbatchnotallowed.yaml
