+++ cmd/operator-opamp-bridge/agent/testdata/agentnoprocessorsallowed.yaml
