+++ cmd/operator-opamp-bridge/agent/testdata/basic.yaml
