+++ cmd/operator-opamp-bridge/agent/testdata/invalid.yaml
