+++ cmd/operator-opamp-bridge/agent/testdata/updated.yaml
