+++ cmd/operator-opamp-bridge/operator/client_test.go
