+++ cmd/operator-opamp-bridge/operator/testdata/collector.yaml
