+++ cmd/operator-opamp-bridge/operator/testdata/invalid-collector.yaml
