+++ cmd/operator-opamp-bridge/operator/testdata/updated-collector.yaml
